Fixed HUD
Moved HUD images to other package

+++ src/main/java/rpg/Statics.java
@@ -9,7 +9,7 @@ public class Statics {
 	private static final double scaleConst = 1.75d;
 	public static final double scale = frameSize.width / 1920d * scaleConst;
 	
-	public static int scale(final int value) {
+	public static int scale(final double value) {
 		return (int) Math.round(value * scale);
 	}
 }
